requirements.txt
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
crcmod==1.7
cryptography==35.0.0
decorator==5.1.0
Flask==2.0.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.1.0
google-cloud-datastore==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oss2==2.15.0
proto-plus==1.19.7
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
PyYAML==6.0
requests==2.26.0
retry==0.9.2
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
